+++ app/controllers/madmin/base_controller.rb
@@ -1,5 +1,6 @@
 module Madmin
   class BaseController < ActionController::Base
+    include ::ActiveStorage::SetCurrent if defined?(::ActiveStorage)
     include Pagy::Backend
 
     protect_from_forgery with: :exception