+++ app/helpers/madmin/application_helper.rb
@@ -1,18 +1,7 @@
 module Madmin
   module ApplicationHelper
     include Pagy::Frontend
-
-    # Converts a Rails version to a NPM version
-    def npm_rails_version
-      version = [
-        Rails::VERSION::MAJOR,
-        Rails::VERSION::MINOR,
-        Rails::VERSION::TINY
-      ].join(".")
-
-      version += "-#{Rails::VERSION::PRE}" if Rails::VERSION::PRE
-      version
-    end
+    include Rails.application.routes.url_helpers
 
     def clear_search_params
       resource.index_path(sort: params[:sort], direction: params[:direction])