+++ app/helpers/madmin/sort_helper.rb
@@ -8,7 +8,7 @@ module Madmin
         concat title
         if matching_column
           concat " "
-          concat tag.i((sort_direction == "asc") ? "▲" : "▼")
+          concat tag.span((sort_direction == "asc") ? asc_icon : desc_icon)
         end
       end
     end
@@ -27,6 +27,22 @@ module Madmin
 
     def default_sort_direction
       resource.try(:default_sort_direction) || "desc"
+    end
+
+    def asc_icon
+      <<~SVG.html_safe
+        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" fill="currentColor" height="1rem" width="1rem">
+          <path fill-rule="evenodd" d="M11.78 9.78a.75.75 0 0 1-1.06 0L8 7.06 5.28 9.78a.75.75 0 0 1-1.06-1.06l3.25-3.25a.75.75 0 0 1 1.06 0l3.25 3.25a.75.75 0 0 1 0 1.06Z" clip-rule="evenodd" />
+        </svg>
+      SVG
+    end
+
+    def desc_icon
+      <<~SVG.html_safe
+        <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 16 16" fill="currentColor" height="1rem" width="1rem">
+          <path fill-rule="evenodd" d="M4.22 6.22a.75.75 0 0 1 1.06 0L8 8.94l2.72-2.72a.75.75 0 1 1 1.06 1.06l-3.25 3.25a.75.75 0 0 1-1.06 0L4.22 7.28a.75.75 0 0 1 0-1.06Z" clip-rule="evenodd" />
+        </svg>
+      SVG
     end
   end
 end