+++ app/javascript/madmin/application.js
@@ -1,0 +1,4 @@
+import "@hotwired/turbo-rails"
+import "trix"
+import "@rails/actiontext"
+import "controllers"