+++ app/javascript/madmin/controllers/application.js
@@ -1,0 +1,12 @@
+import { Application } from "@hotwired/stimulus"
+
+const application = Application.start()
+
+// Configure Stimulus development experience
+application.debug = false
+window.Stimulus   = application
+
+export { application }
+
+import { Dropdown } from "tailwindcss-stimulus-components"
+application.register("dropdown", Dropdown)