+++ lib/madmin/engine.rb
@@ -1,3 +1,5 @@
+require "importmap-rails"
+
 module Madmin
   class Engine < ::Rails::Engine
     isolate_namespace Madmin
@@ -9,10 +11,30 @@ module Madmin
 
     config.to_prepare do
       Madmin.reset_resources!
+      Madmin.site_name ||= Rails.application.class.module_parent_name
     end
 
     initializer "madmin.assets" do |app|
-      app.config.assets.precompile += %w[madmin_manifest]
+      if app.config.respond_to?(:assets)
+        app.config.assets.paths << root.join("app/assets/stylesheets")
+        app.config.assets.paths << root.join("app/javascript")
+        app.config.assets.precompile += %w[madmin_manifest]
+
+        Madmin.stylesheets << if defined?(::Sprockets)
+          "madmin/application-sprockets"
+        else
+          "madmin/application"
+        end
+      end
+    end
+
+    initializer "madmin.importmap", before: "importmap" do |app|
+      Madmin.importmap.draw root.join("config/importmap.rb")
+      Madmin.importmap.cache_sweeper watches: root.join("app/javascript")
+
+      ActiveSupport.on_load(:action_controller_base) do
+        before_action { Madmin.importmap.cache_sweeper.execute_if_updated }
+      end
     end
   end
 end