+++ lib/madmin/fields/select.rb
@@ -1,0 +1,9 @@
+module Madmin
+  module Fields
+    class Select < Field
+      def options_for_select(record)
+        options.collection
+      end
+    end
+  end
+end