+++ test/dummy/app/models/post.rb
@@ -1,7 +1,6 @@
 class Post < ApplicationRecord
   extend FriendlyId
   friendly_id :title
-  has_paper_trail
 
   belongs_to :user
   has_many :comments, as: :commentable, dependent: :destroy
@@ -11,7 +10,7 @@ class Post < ApplicationRecord
 
   scope :recent, -> { where(created_at: 2.weeks.ago..) }
 
-  enum state: [:draft, :published, :archived]
+  enum :state, [:draft, :published, :archived]
 
   validates :title, presence: true
 end