+++ test/dummy/config/initializers/dummy.rb
@@ -1,0 +1,1 @@
+Madmin.stylesheets << "dummy"